refactor(Face): separate file for Face struct
